+++ packages/core/src/material/index.ts
@@ -1,11 +1,11 @@
-export { RenderQueueType } from "./enums/RenderQueueType";
-export { RenderFace } from "./enums/RenderFace";
-export { BlendMode } from "./enums/BlendMode";
-export { TextureCoordinate } from "./enums/TextureCoordinate";
-export { Material } from "./Material";
 export { BaseMaterial } from "./BaseMaterial";
 export { BlinnPhongMaterial } from "./BlinnPhongMaterial";
+export { BlendMode } from "./enums/BlendMode";
+export { RenderFace } from "./enums/RenderFace";
+export { TextureCoordinate } from "./enums/TextureCoordinate";
+export { Material } from "./Material";
 export { PBRBaseMaterial } from "./PBRBaseMaterial";
 export { PBRMaterial } from "./PBRMaterial";
 export { PBRSpecularMaterial } from "./PBRSpecularMaterial";
 export { UnlitMaterial } from "./UnlitMaterial";
+
