+++ packages/core/src/shader/index.ts
@@ -1,3 +1,4 @@
+export { RenderQueueType } from "./enums/RenderQueueType";
 export { BlendFactor } from "./enums/BlendFactor";
 export { BlendOperation } from "./enums/BlendOperation";
 export { ColorWriteMask } from "./enums/ColorWriteMask";
@@ -5,5 +6,6 @@ export { CompareFunction } from "./enums/CompareFunction";
 export { CullMode } from "./enums/CullMode";
 export { StencilOperation } from "./enums/StencilOperation";
 export { Shader } from "./Shader";
-export { ShaderPass } from "./ShaderPass";
 export { ShaderData } from "./ShaderData";
+export { ShaderPass } from "./ShaderPass";
+
